@@ -15,6 +15,7 @@
     "start": "node server.js",
     "start:useDist": "DEV_USE_DIST=yes node server.js"
   },
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/lyrgard/ffbeEquip.git"
@@ -26,50 +27,49 @@
   },
   "homepage": "https://github.com/lyrgard/ffbeEquip#readme",
   "dependencies": {
-    "@sindresorhus/is": "0.11.0",
-    "ajv": "^6.5.3",
-    "bluebird": "3.5.1",
-    "body-parser": "^1.18.3",
-    "boom": "7.2.0",
+    "@hapi/boom": "^10.0.0",
+    "@sindresorhus/is": "5.3.0",
+    "ajv": "^8.11.0",
+    "bluebird": "3.7.2",
+    "body-parser": "^1.20.0",
     "client-sessions": "0.8.0",
     "cors": "^2.8.5",
-    "crypto-random-string": "1.0.0",
-    "express": "^4.16.3",
-    "firebase-admin": "^5.12.0",
-    "googleapis": "26.0.1",
-    "grpc": "^1.24.11",
-    "helmet": "^3.13.0",
-    "inquirer": "6.2.0",
-    "joi": "17.2.1",
-    "morgan": "1.9.0",
-    "pngjs": "3.3.2",
+    "crypto-random-string": "5.0.0",
+    "es-main": "^1.2.0",
+    "express": "^4.18.1",
+    "firebase-admin": "^11.0.1",
+    "googleapis": "105.0.0",
+    "helmet": "^5.1.1",
+    "inquirer": "9.1.0",
+    "joi": "17.6.0",
+    "lodash": "^4.17.21",
+    "pngjs": "6.0.0",
     "request": "^2.87.0",
-    "uuid": "^3.3.2"
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "babel-loader": "^8.0.0",
-    "babel-preset-minify": "^0.4.3",
-    "clean-webpack-plugin": "^0.1.19",
-    "cssnano": "4.0.5",
-    "exports-loader": "0.7.0",
-    "extract-loader": "2.0.1",
-    "file-loader": "2.0.0",
-    "glob": "7.1.2",
-    "html-loader": "0.5.5",
-    "image-size": "^0.6.3",
-    "md5-file": "^4.0.0",
-    "mocha": "5.2.0",
-    "mocha-webpack": "2.0.0-beta.0",
-    "mock-require": "3.0.2",
-    "nock": "9.6.1",
-    "nyc": "12.0.2",
-    "postcss-loader": "3.0.0",
+    "@babel/core": "^7.18.10",
+    "babel-loader": "^8.2.5",
+    "babel-preset-minify": "^0.5.2",
+    "clean-webpack-plugin": "^4.0.0",
+    "cssnano": "5.1.12",
+    "exports-loader": "4.0.0",
+    "extract-loader": "5.1.0",
+    "file-loader": "6.2.0",
+    "glob": "8.0.3",
+    "html-loader": "4.1.0",
+    "image-size": "^1.0.2",
+    "md5-file": "^5.0.0",
+    "mocha": "10.0.0",
+    "mock-require": "3.0.3",
+    "nock": "13.2.9",
+    "nyc": "15.1.0",
+    "postcss-loader": "7.0.1",
     "string-replace-webpack-plugin": "^0.1.3",
-    "supertest": "3.1.0",
-    "webpack": "4.17.1",
-    "webpack-cli": "3.1.0",
-    "webpack-node-externals": "1.7.2"
+    "supertest": "6.2.4",
+    "webpack": "5.74.0",
+    "webpack-cli": "4.10.0",
+    "webpack-node-externals": "3.0.0"
   },
   "nyc": {
     "all": true,
