@@ -1,10 +1,12 @@
-const Boom = require('boom');
+import Boom from '@hapi/boom'
 
 /* eslint no-unused-vars: 0 */
-module.exports = (err, req, res, next) => {
+export function boomJS(err, req, res, next) {
   const error = (err.isBoom) ? err : Boom.boomify(err);
 
   return res
     .status(error.output.statusCode)
     .json(error.output.payload);
 };
+
+export default { boomJS }
