@@ -1,6 +1,6 @@
-const OAuth = require('../lib/oauth.js');
+import OAuth  from "../lib/oauth.js";
 
-module.exports = (req, res, next) => {
+export function OAuthFunction(req, res, next){
   const { tokens } = req.OAuthSession;
   if (!tokens) {
     return res.status(401).send();
@@ -8,8 +8,10 @@ module.exports = (req, res, next) => {
 
   req.OAuth2Client = OAuth.createClient(tokens);
 
-  return req.OAuth2Client.getRequestMetadata(null, (error) => {
+  return req.OAuth2Client.getRequestMetadataAsync(null, (error) => {
     req.OAuthSession.tokens = req.OAuth2Client.credentials;
     next(error);
   });
 };
+
+export default { OAuthFunction }
