@@ -1,7 +1,6 @@
-const config = require('../../config.js');
-const express = require('express');
+import express from 'express'
 
-const route = express.Router();
+export const route = express.Router();
 
 /**
  * "GET /"
@@ -12,4 +11,3 @@ route.get('/', (req, res) => {
   return res.send("window.clientConfig = {firebaseBucketUri:'" + config.firebase.bucketUri + "', firebaseDatabaseId:'" + config.firebase.databaseId + "', imgurClientId:'" + config.imgur.clientId + "'}");
 });
 
-module.exports = route;
