@@ -1607,8 +1607,8 @@ function prepareSearch(data) {
         if (item["exclusiveSex"]) {
             textToSearch += "|Only " + item["exclusiveSex"];
         }
-        if (item.exclusiveRoles) {
-            textToSearch += "|Only " + item.exclusiveRoles.join(', ');
+        if (item["exclusiveRoles"]) {
+            textToSearch += "|Only " + item["exclusiveRoles"].join(', ');
         }
         if (item["condition"]) {
             textToSearch += "|Only " + item["condition"];
@@ -1784,6 +1784,7 @@ function getShortUrl(longUrl, callback) {
 }
 
 function getLocalizedFileUrl(name) {
+    console.log("Localized")
     if (language) {
         name = name + "_" + language;
     }
@@ -1792,12 +1793,19 @@ function getLocalizedFileUrl(name) {
 }
 
 function onUnitsOrInventoryLoaded() {
+    console.log("OnUnits")
+    console.log(itemInventory);
+    console.log(ownedUnits);
+    console.log(ownedEspers);
+    console.log(ownedConsumables);
     if (itemInventory && ownedUnits && ownedEspers && ownedConsumables) {
         if (ownedUnits.version && ownedUnits.version < 3) {
             // before version 3, units were : {"unitId": number}
             // After, they are {"unitId": {"number":number,"farmable":number}
             $.get(getLocalizedFileUrl("data"), function(data) {
-                $.get(server + "/units.json", function(unitResult) {
+                console.log("Here Localized")
+                $.get("/" + server + "/units.json", function(unitResult) {
+                    console.log("here")
                     var allUnitsTmp = unitResult;
                     var tmrNumberByUnitId = {};
                     for (var index = data.length; index--; ) {
@@ -2500,6 +2508,7 @@ $(function() {
     try {
         // Bust the whole localStorage in case of old array used in order to get a clean state
         // @TODO: can be removed after october 2018
+        console.log(localStorages)
         if (localStorageAvailable && $.isArray(JSON.parse(localStorage.getItem("savedFiles")))) {
             localStorage.clear();
             window.console && window.console.warn("Clearing the whole localStorage!");
@@ -2509,7 +2518,7 @@ $(function() {
 
     readUrlParams();
 
-    $.get(server + '/dataVersion.json', function(result) {
+    $.get("/" + server + '/dataVersion.json', function(result) {
         var dataVersion = result.version;
         var selectedLanguage = language ? language : "en";
 
@@ -2534,7 +2543,10 @@ $(function() {
         $("#inventoryDiv").removeClass("Inventoryloading Inventoryloaded");
         notLoaded();
     } else {
-        $.get(server + '/itemInventory', function(result) {
+        console.log("Getting Item Inventory...")
+
+        $.get("/" + server + '/itemInventory', function(result) {
+            console.log("Got inventory")
             itemInventory = result;
             if (!itemInventory.enchantments) {
                 itemInventory.enchantments = {};
@@ -2545,16 +2557,18 @@ $(function() {
             adaptItemInventoryForMultipleRareEnchantments();
             sanitizeItemInventory();
             onUnitsOrInventoryLoaded();
-        }, 'json').fail(function(jqXHR, textStatus, errorThrown ) {
+        }, 'json')
+        .fail(function(jqXHR, textStatus, errorThrown ) {
+            console.log("File is not available. Please log in to gain access to inventory data.")
             $("#inventoryDiv").removeClass("Inventoryloading Inventoryloaded");
             if (notLoaded) {
                 notLoaded();
             }
         });
-        $.get(server + '/settings', function(result) {
+        $.get("/" + server + '/settings', function(result) {
             userSettings = result;
         });
-        $.get(server + '/units', function(result) {
+        $.get("/" + server + '/units', function(result) {
             ownedUnits = result;
             if (result.version && result.version == 3) {
                 getStaticData("units", false, function(allUnitResult) {
@@ -2589,7 +2603,7 @@ $(function() {
             }
         });
         console.log("Starts to load owned espers");
-        $.get(server + '/espers', function(result) {
+        $.get("/" + server + '/espers', function(result) {
             ownedEspers = result;
 
             Object.keys(ownedEspers).forEach(esper => {
@@ -2608,7 +2622,7 @@ $(function() {
             }
         });
         console.log("Starts to load owned consumables");
-        $.get(server + '/consumables', function(result) {
+        $.get("/" + server + '/consumables', function(result) {
             ownedConsumables = result;
             console.log("owned consumables loaded");
             onUnitsOrInventoryLoaded();
